automation/requirements.txt
arxiv>=2.1.3
httpx>=0.27.0
pydantic>=2.9,<3
pydantic-settings>=2.5,<3
supabase>=2.10.0
python-dotenv>=1.0.1
